requirements.txt
Flask==2.3.5
requests==2.31.0
numpy==1.26.0
pandas==2.1.1
openai==1.8.0
tiktoken==0.4.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==8.3.0
